refactor: group type guards in one directory
